Job Hunter/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
discord.py==2.6.4
feedparser==6.0.12
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
pillow==12.0.0
propcache==0.4.1
python-dotenv==1.2.1
requests==2.32.5
sgmllib3k==1.0.0
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
